move alignment code to separate file
