+++ Makefile.am
@@ -72,6 +72,7 @@ include_HEADERS    = ./alp/gumbel_params.hpp \
                      ./include/load_index.hpp \
                      ./include/traverse_bursttrie.hpp \
                      ./include/mmap.hpp \
+                     ./include/alignment.hpp \
                      ./include/paralleltraversal.hpp \
                      ./include/ssw.h \
                      ./include/outputformats.hpp
@@ -119,6 +120,7 @@ sortmerna_SOURCES  = ./alp/gumbel_params.cpp \
                      ./src/load_index.cpp \
                      ./src/traverse_bursttrie.cpp \
                      ./src/mmap.cpp \
+                     ./src/alignment.cpp \
                      ./src/paralleltraversal.cpp \
                      ./src/ssw.c \
                      ./src/outputformats.cpp \
